Move ArrayCow method impls into separate module
This is analogous to `impl_owned_array`, `impl_views`, and
`impl_raw_views`.
